refactor: 827 - rename mentorship file
