+++ src/views/mentorship/mentorship.tsx
@@ -1,12 +1,12 @@
-import { MentorActivities } from '../../widgets/mentorship-activities';
-import { MentorshipCourses } from '../../widgets/mentorship-courses/mentorship-courses';
-import { DetailsMentorship } from '../../widgets/mentorship-details/details/details';
-import { MentorshipHero } from '../../widgets/mentorship-hero';
 import { Benefits } from '@/widgets/benefits';
-import { MentorsAfterRegister } from '@/widgets/mentors-after-register';
+import { MentorshipAfterRegister } from '@/widgets/mentors-after-register';
 import { MentorsDocs } from '@/widgets/mentors-docs';
 import { MentorsFeedback } from '@/widgets/mentors-feedback';
 import { MentorsRegister } from '@/widgets/mentors-register';
+import { MentorActivities } from '@/widgets/mentorship-activities';
+import { MentorshipCourses } from '@/widgets/mentorship-courses/mentorship-courses';
+import { DetailsMentorship } from '@/widgets/mentorship-details/details/MentorshipDetails';
+import { MentorshipHero } from '@/widgets/mentorship-hero';
 import { MentorshipCourse, benefitMentorshipMentors, mentorsFeedbackData } from 'data';
 
 type MentorshipProps = {
@@ -27,7 +27,7 @@ export const Mentorship = ({ mentorshipData, courses = false }: MentorshipProps)
       {courses && <MentorshipCourses />}
       <MentorsFeedback mentorsFeedback={mentorsFeedbackData} />
       <MentorsRegister lang={mentorshipData.lang} />
-      <MentorsAfterRegister lang={mentorshipData.lang} />
+      <MentorshipAfterRegister lang={mentorshipData.lang} />
       <MentorsDocs
         mentorDocsLink={mentorshipData.links.mentorDocs}
         courseDocsLink={mentorshipData.links.courseDocs}
