+++ src/widgets/footer/index.ts
