+++ src/widgets/footer/ui/copyright.tsx
