+++ src/widgets/header/header.tsx
@@ -4,8 +4,8 @@ import { useEffect, useState } from 'react';
 import classNames from 'classnames/bind';
 import { usePathname } from 'next/navigation';
 
-import { BurgerMenu } from './burger/burger';
-import { NavItem } from './nav-item/nav-item';
+import { BurgerMenu } from './ui/burger/burger';
+import { NavItem } from './ui/nav-item/nav-item';
 import { ANCHORS, ROUTES } from '@/core/const';
 import { Course } from '@/entities/course';
 import { Logo } from '@/shared/ui/logo';
