+++ src/widgets/partnered/partnered.test.tsx
