+++ src/widgets/partnered/partnered.tsx
